% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{cm2bp}
\alias{cm2bp}
\title{Convert centimorgan to bp}
\usage{
cm2bp(peaks_chr, annot)
}
\arguments{
\item{peaks_chr}{A data.frame with chr, start_cm and end_cm information about
a single chromosome.}

\item{annot}{A data.frame with chr, cm and bp columns.}
}
\value{
The peaks_chr data.frame with the start_bp and end_bp columns inserted.
}
\description{
Converts genetic distance in centimorgans to genetic position in bp.
}
\examples{
par <- partbl("merlin_10_famA-parametric.tbl")
peaks <- peaks(par)
annot <- read_annot("annotHapMap2U.txt")
cm2bp(peaks, annot)

}

