% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famlod.R
\name{famlod}
\alias{famlod}
\title{MERLIN \code{famlod} Object}
\usage{
famlod(fname)
}
\arguments{
\item{fname}{Character vector containing the \code{fam.par} file name(s)
output by MERLIN.}
}
\value{
S3 object of class \code{famlod}, which is a list with:
        \itemize{
          \item famlod: data.frame with 5 columns
               (chr, pos, family, analysis, lod, plod)
          \item max_lods: data.frame with 3 columns
               (chr, family, max_lod, max_plod)
          \item n_markers: data.frame with 2 columns
               (chr, n)
               }
}
\description{
Returns an S3 object of class \code{famlod} which is basically
a compact version of MERLIN's \code{fam.lod} files.
}
\details{
MERLIN's \code{--perFamily} option is used to output LOD scores for each family
based on if parametric and/or nonparametric analyses have been specified.

The nonparametric analysis is specified using
at least one of the options \code{--npl} or \code{--pairs}.
These compute LOD scores using the Kong and Cox linear model
designated in the 'analysis' column of the file by character strings containing
'ALL' and 'pairs', respectively.
The 'plod' score is calculated using the best fitting overall model. The 'lod'
score is maximized within each family (LOD)

For more information, see MERLIN's website:
\url{http://csg.sph.umich.edu/abecasis/merlin/index.html}
}
\examples{
famlod_chr9 <- famlod("merlin_9_famA.lod")
famlod_chr24 <- famlod(c("merlin_2_famA.lod", "merlin_4_famA.lod"))
famlod_all <- famlod(list.files(path = "merlin/", pattern = "lod$"))

}

