% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_famlod.R
\name{famlod2ggdata}
\alias{famlod2ggdata}
\title{\code{famlod} GG Data Preparation}
\usage{
famlod2ggdata(famlod)
}
\arguments{
\item{famlod}{S3 object of class \code{famlod}.}
}
\value{
A data.frame with 6 columns: chr (num), pos (num), family (chr),
        analysis (chr), variable (chr) and value (num).
}
\description{
\code{famlod} GG Data Preparation
}
\examples{
fl <- famlod("foo.lod")
fl_ggdata <- famlod2ggdata(pt)
head(fl_ggdata)

}

