% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fampar.R
\name{fampar}
\alias{fampar}
\title{MERLIN \code{fampar} Object}
\usage{
fampar(fname)
}
\arguments{
\item{fname}{Character vector containing the \code{fam.par} file name(s)
output by MERLIN.}
}
\value{
S3 object of class \code{partbl}, which is a list with:
        \itemize{
          \item fampar: data.frame with 5 columns
               (chr, model, family, position, lod)
          \item max_lods: data.frame with 3 columns
               (chr, family, max_lod)
          \item n_markers: data.frame with 2 columns
               (chr, n)
               }
}
\description{
Returns an S3 object of class \code{fampar} which is basically
a compact version of MERLIN's \code{fam.par} files.
}
\details{
MERLIN's \code{--perFamily} option is used to output LOD scores for each family
based on if parametric and/or nonparametric analyses have been specified.

The parametric analysis is specified using the option
\code{--model genetic_model.txt}.
For more information, see MERLIN's website:
\url{http://csg.sph.umich.edu/abecasis/merlin/index.html}
}
\examples{
fampar_chr9 <- fampar("merlin_9_famA.par")
fampar_chr24 <- fampar(c("merlin_2_famA.par", "merlin_4_famA.par"))
fampar_all <- fampar(list.files(path = "merlin/", pattern = "par$"))

}

