% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npartbl.R
\name{get_npartbl}
\alias{get_npartbl}
\title{Get \code{npartbl} Table}
\usage{
get_npartbl(npartbl)
}
\arguments{
\item{npartbl}{An object of class \code{npartbl}.}
}
\value{
A data.frame with the chromosome, pos_cm, analysis,
lod and exlod columns from a \code{npartbl} object.
}
\description{
Get \code{npartbl} Table
}
\examples{
get_npartbl(npartbl("fam_nonparametric.tbl"))

}

