% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{get_partbl}
\alias{get_partbl}
\title{Get \code{partbl} Table}
\usage{
get_partbl(partbl)
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}
}
\value{
A data.frame with the chromosome, position, model,
lod, alpha and hlod columns from a \code{partbl} object.
}
\description{
Get \code{partbl} Table
}
\examples{
get_partbl(partbl("fam_parametric.tbl"))

}

