% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkrvis.R
\docType{package}
\name{linkrvis}
\alias{linkrvis}
\alias{linkrvis-package}
\title{linkrvis: A package for visualising output from genetic linkage analysis tools}
\description{
linkrvis: A package for visualising output from genetic linkage analysis tools
}

