% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_vec}
\alias{match_vec}
\title{Match Values to Given Vector}
\usage{
match_vec(x, a = "alpha", b = "beta")
}
\arguments{
\item{x}{A character vector of length 1 or 2}

\item{a}{A character vector of length 1}

\item{b}{A character vector of length 1}
}
\value{
A character vector of length 1 or 2, containing
the string in variable a, b or both.
}
\description{
Match Values to Given Vector
}
\examples{
match_vec(c("beta", "alpha"))
match_vec("alpha")
match_vec("betaaaa")
match_vec("bettaaaa")
match_vec("alphaandbeta")
match_vec(c("this has thealpha", "beta in here"))

}

