% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_npartbl.R
\name{npartbl2ggdata}
\alias{npartbl2ggdata}
\title{\code{npartbl} GG Data Preparation}
\usage{
npartbl2ggdata(npartbl, vars = c("lod", "exlod"))
}
\arguments{
\item{npartbl}{S3 object of class \code{partbl}.}

\item{vars}{A character vector indicating which variable(s) to plot. Available
variables from the \code{partbl} file are: "lod", "alpha" and "hlod".}
}
\value{
A data.frame with three columns: pos (num), analysis (chr),
        variable (chr) and value (num).
}
\description{
\code{npartbl} GG Data Preparation
}
\examples{
npt <- npartbl("foo-nparametric.tbl")
npt_ggdata <- npartbl2ggdata(pt)
head(npt_ggdata)

}

