% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merlin.R
\name{partbl2ggdata}
\alias{partbl2ggdata}
\title{\code{partbl} GG Data Preparation}
\usage{
partbl2ggdata(partbl, vars = c("lod", "alpha", "hlod"), threshold = 0)
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}

\item{vars}{A character vector indicating which variable(s) to plot. Available
variables from the \code{partbl} file are: "lod", "alpha" and "hlod".}
}
\value{
A data.frame with three columns: pos (num), variable (chr) and value (num).
}
\description{
\code{partbl} GG Data Preparation
}
\examples{
pt <- partbl("foo-parametric.tbl")
pt_ggdata <- partbl2ggdata(pt)
head(pt_ggdata)

}

