% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_famlod.R
\name{plot.famlod}
\alias{plot.famlod}
\title{Plot MERLIN PerFam Nonparametric LOD Scores}
\usage{
\method{plot}{famlod}(famlod, ylim = c(0, max(famlod$max_lods[c("max_lod",
  "max_plod")])))
}
\arguments{
\item{famlod}{S3 object of class \code{famlod}.}

\item{ylim}{Character vector of length 2 indicating the range of values to plot.}
}
\value{
ggplot object, plotting the selected variables
}
\description{
Plots the LOD scores in MERLIN's \code{fam.lod} file
}
\examples{
fl <- famlod("merlin_10_famA.lod")
plot(fl)

}

