% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fampar.R
\name{plot.fampar}
\alias{plot.fampar}
\title{Plot MERLIN PerFam Parametric LOD Scores}
\usage{
\method{plot}{fampar}(fampar, ylim = c(0, max(fampar$max_lods$max_lod)))
}
\arguments{
\item{fampar}{S3 object of class \code{fampar}.}

\item{ylim}{Character vector of length 2 indicating the range of values to plot.}
}
\value{
ggplot object, plotting the selected variables
}
\description{
Plots the LOD scores in MERLIN's \code{fam.par} file
}
\examples{
fp <- fampar("merlin_10_famA.par")
plot(fp)

}

