% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_npartbl.R
\name{plot.npartbl}
\alias{plot.npartbl}
\title{Plot MERLIN Nonparametric LOD Scores}
\usage{
\method{plot}{npartbl}(npartbl, vars = c("lod", "exlod"), ylim = c(0,
  max(npartbl$max_lods[c("max_lod", "max_exlod")])))
}
\arguments{
\item{npartbl}{S3 object of class \code{npartbl}.}

\item{vars}{Character vector indicating which variable(s) to plot. Available
variables from the \code{npartbl} data.frame are: "lod" and "exlod".}

\item{ylim}{Character vector of length 2 indicating the range of values to plot.}
}
\value{
ggplot object, plotting the selected variables
}
\description{
Plots the LOD and ExLOD scores in MERLIN's \code{nonparametric.tbl} file
}
\examples{
npt <- npartbl("foo-nonparametric.tbl")
plot(npt)

}

