% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merlin.R
\name{plot.partbl}
\alias{plot.partbl}
\title{Plot MERLIN Parametric LOD Scores}
\usage{
\method{plot}{partbl}(partbl, ...)
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}

\item{...}{Other arguments passed on to \code{\link{partbl2ggdata}}.}
}
\value{
A ggplot object, plotting the selected variables
}
\description{
Plots the LOD, ALPHA and HLOD scores in MERLIN's \code{famA-parametric.tbl} file
}
\examples{
pt <- partbl("foo-parametric.tbl")
plot(pt)

}

