% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famlod.R
\name{print.famlod}
\alias{print.famlod}
\title{Print Head and Tail of \code{famlod} Table}
\usage{
\method{print}{famlod}(famlod, n = 6L)
}
\arguments{
\item{famlod}{An object of class \code{famlod}.}

\item{n}{Number of rows to show in the head and tail}
}
\description{
Prints the head and tail of a \code{famlod} table.
}

