% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fampar.R
\name{print.fampar}
\alias{print.fampar}
\title{Print Head and Tail of \code{fampar} Table}
\usage{
\method{print}{fampar}(fampar, n = 6L)
}
\arguments{
\item{fampar}{An object of class \code{fampar}.}

\item{n}{Number of rows to show in the head and tail}
}
\description{
Prints the head and tail of a \code{fampar} table.
}

