% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npartbl.R
\name{print.npartbl}
\alias{print.npartbl}
\title{Print Head, Middle and Tail of \code{npartbl} Table}
\usage{
\method{print}{npartbl}(npartbl, n = 6L)
}
\arguments{
\item{npartbl}{An object of class \code{npartbl}.}

\item{n}{Number of rows to show in the head, middle and tail}
}
\description{
Prints the head, middle and tail of a \code{npartbl} table.
}

