% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{print.partbl}
\alias{print.partbl}
\title{Print Head and Tail of \code{partbl} Table}
\usage{
\method{print}{partbl}(partbl, n = 6L)
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}

\item{n}{Number of rows to show in the head and tail}
}
\description{
Prints the head and tail of a \code{partbl} table.
}

