% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famlod.R
\name{print.summary.famlod}
\alias{print.summary.famlod}
\title{Print Summary of \code{famlod} Object}
\usage{
\method{print}{summary.famlod}(famlod_summary)
}
\arguments{
\item{famlod_summary}{An object of class \code{summary.famlod}}
}
\description{
Print Summary of \code{famlod} Object
}

