% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fampar.R
\name{print.summary.fampar}
\alias{print.summary.fampar}
\title{Print Summary of \code{fampar} Object}
\usage{
\method{print}{summary.fampar}(fampar_summary)
}
\arguments{
\item{fampar_summary}{An object of class \code{summary.fampar}}
}
\description{
Print Summary of \code{fampar} Object
}

