% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{print.summary.partbl}
\alias{print.summary.partbl}
\title{Print Summary of \code{partbl} Object}
\usage{
\method{print}{summary.partbl}(partbl_summary)
}
\arguments{
\item{partbl}{An object of class \code{summary.partbl}}
}
\description{
Print Summary of \code{partbl} Object
}

