% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merlin.R
\name{read_annot}
\alias{read_annot}
\title{Read HapMap2 Annotation File}
\usage{
read_annot(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
data.frame with 4,031,388 rows and 3 columns: chrom (char), bp (int)
        and cm (double)
}
\description{
Read the LINKDATAGEN annotation file for MPS and Illumina SNP chip data
(rev.807 Dec2013), and select only the chr, bp and cm columns.
Based on the b37 build.
}
\examples{
annot <- read_annot("annotHapMap2U.txt.gz")

}

