% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merlin.R
\name{read_merlin_dat}
\alias{read_merlin_dat}
\title{Read MERLIN \code{fam.dat}}
\usage{
read_merlin_dat(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{data.frame} with two columns (symbol and description) and M + 1
        rows, where M is the number of markers used in the analysis and the last
        row describes the phenotype of interest.
}
\description{
Read the \code{fam.dat} file generated by LDG for MERLIN
}
\details{
The \code{fam.dat} file contains two columns: symbol and description.
This file is required to describe the final M + 1 columns in the \code{fam.ped}
file, where M is the number of markers used in the analysis and the last column
is the affectedness status (the first five columns are always
the same - FamID, IndID, PatID, MatID and Gender). Therefore the \code{fam.dat}
file will contain M + 1 rows.
}
\examples{
read_merlin_dat("merlin_13_famA.dat")
}

