% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fampar.R
\name{read_merlin_fampar}
\alias{read_merlin_fampar}
\title{Read MERLIN \code{fam.par}}
\usage{
read_merlin_fampar(fname, chrom = as.numeric(sub(".*merlin_(\\\\d{1,2})_.*",
  "\\\\1", fname)), verbose = TRUE)
}
\arguments{
\item{fname}{The file name to read.}

\item{chrom}{The chromosome number for this file. This is the number in
the filename following "merlin_" e.g. for "merlin_13_famA.par" it will
be 13.}

\item{verbose}{}
}
\value{
A \code{data.frame} with the columns
}
\description{
Read the \code{.par} file output by MERLIN.
}
\examples{
read_merlin_fampar("merlin_13_famA.par")
}

