% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merlin.R
\name{read_merlin_map}
\alias{read_merlin_map}
\title{Read MERLIN \code{fam.map}}
\usage{
read_merlin_map(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{data.frame} with three columns (chromosome, name and position)
and M rows, where M is the number of markers used in the analysis.
}
\description{
Read the \code{fam.map} file generated by LDG for MERLIN
}
\details{
The \code{fam.map} file contains three columns: chromosome, marker name, and
marker position in centimorgans.
}
\examples{
read_merlin_map("merlin_13_famA.map")
}

