% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merlin.R
\name{read_merlin_npartbl}
\alias{read_merlin_npartbl}
\title{Read MERLIN \code{fam_nonparametric.tbl}}
\usage{
read_merlin_npartbl(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{data.frame}.
}
\description{
Read the \code{fam_nonparametric.tbl} file output by MERLIN
}
\details{
MERLIN's \code{--tabulate} option outputs convenient tables summarising
linkage analysis results in two tables:
\code{fam_nonparametric.tbl} and \code{fam_parametric.tbl}.

The \code{fam_nonparametric.tbl} file will be useful if MERLIN was run with
at least one of the options \code{--npl}, \code{--pairs} or \code{--exp}.
The first two compute a LOD score using the Kong and Cox linear model, and are
designated in the file as 'all' and 'pairs', respectively;
the third complements the LOD score by computing an ExLOD score using the
Kong and Cox exponential mode.

The \code{fam_nonparametric.tbl} file contains two rows at the
beginning of each \code{all} and/or \code{pairs} subtable, which
indicate the maximum possible scores for the specific dataset.
These rows are ignored by using the comment.char = "n"
option in \code{\link{read.table}}, which works since only these rows
contain an 'n' character. This is a bit hacky, but it should always work.
}
\examples{
read_merlin_npartbl("merlin_13_famA-nonparametric.tbl")
}

