% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merlin.R
\name{read_merlin_ped}
\alias{read_merlin_ped}
\title{Read MERLIN \code{fam.ped}}
\usage{
read_merlin_ped(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{matrix} with N rows and 5 + M + 1 columns
(refer to the details section).
}
\description{
Read the \code{fam.ped} file generated by LDG for MERLIN
}
\details{
The \code{fam.ped} file contains N rows and 5 + M + 1 columns, where N is
the number of samples, M is the number of markers and the last column
is the affectedness status (the first five columns are always
the same - FamID, IndID, PatID, MatID and Gender).
}
\examples{
read_merlin_ped("merlin_13_famA.ped")
}

