% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_npartbl}
\alias{read_npartbl}
\title{Read MERLIN \code{fam_parametric.tbl}}
\usage{
read_npartbl(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{data.frame}.
}
\description{
Read the \code{fam_nonparametric.tbl} file output by MERLIN
}
\details{
The \code{fam_nonparametric.tbl} file contains two rows at the
beginning of each \code{ALL} and/or \code{Pairs} subtable, which
indicate the maximum possible scores for the specific dataset.
These rows are ignored by using the comment.char = "n"
option in \code{\link{read.table}}, which works since only these rows
contain an 'n' character. This is a bit hacky, but it should always work.
}
\examples{
read_npartbl("merlin_13_famA-nonparametric.tbl")
}

