% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_partbl}
\alias{read_partbl}
\title{Read MERLIN \code{fam_parametric.tbl}}
\usage{
read_partbl(fname)
}
\arguments{
\item{fname}{The file name to read.}
}
\value{
A \code{data.frame}.
}
\description{
Read the \code{fam_parametric.tbl} file output by MERLIN.
}
\examples{
read_partbl("merlin_13_famA-parametric.tbl")
}

