% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{summarise_peaks}
\alias{summarise_peaks}
\title{Summarise LOD score peaks}
\usage{
summarise_peaks(ld, valcol)
}
\arguments{
\item{ld}{A list of data.frames containing markers achieving high LOD scores.}

\item{valcol}{A length-one character vector giving the name of the column in
\code{d} containing the values.}
}
\value{
A data.frame with chr, start_cm, end_cm, etc.
}
\description{
Which regions achieve a LOD score above a given threshold?
}
\examples{
partbl <- partbl::partbl("merlin_10_famA-parametric.tbl")
list_of_peaks <- get_peaks(partbl)
summarise_peaks(list_of_peaks)

}

