% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famlod.R
\name{summary.famlod}
\alias{summary.famlod}
\title{Return Summary of \code{famlod} Object}
\usage{
\method{summary}{famlod}(famlod)
}
\arguments{
\item{famlod}{An object of class \code{famlod}.}
}
\value{
A list (of class \code{summary.famlod}) containing summary
information about the \code{famlod} object.
}
\description{
Return Summary of \code{famlod} Object
}

