% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fampar.R
\name{summary.fampar}
\alias{summary.fampar}
\title{Return Summary of \code{fampar} Object}
\usage{
\method{summary}{fampar}(fampar)
}
\arguments{
\item{fampar}{An object of class \code{fampar}.}
}
\value{
A list (of class \code{summary.fampar}) containing summary
information about the \code{fampar} object.
}
\description{
Return Summary of \code{fampar} Object
}

