% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npartbl.R
\name{summary.npartbl}
\alias{summary.npartbl}
\title{Return Summary of \code{npartbl} Object}
\usage{
\method{summary}{npartbl}(npartbl)
}
\arguments{
\item{npartbl}{An object of class \code{npartbl}.}
}
\value{
A list (of class \code{summary.npartbl}) containing summary
information about the \code{npartbl} object.
}
\description{
Return Summary of \code{npartbl} Object
}

