% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{summary.partbl}
\alias{summary.partbl}
\title{Return Summary of \code{partbl} Object}
\usage{
\method{summary}{partbl}(partbl)
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}
}
\value{
A list (of class \code{summary.partbl}) containing summary
information about the \code{partbl} object.
}
\description{
Return Summary of \code{partbl} Object
}

