% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Output Rows with Values Above Threshold}
\usage{
get_peaks(d, valcol = "lod", t = 0)
}
\arguments{
\item{d}{An object of class \code{\link{partbl}}, \code{link{npartbl}},
\code{link{famlod}} or \code{link{fampar}}.
Can also be a data.frame with two numeric columns
that make sense as 'position' and 'value'.}

\item{valcol}{A length-one character vector giving the name of the column in
\code{d} containing the values.}

\item{t}{A length-one numeric vector specifying the threshold to be used.}
}
\value{
A list where each element is a data.frame containing subsets of \code{d}
        where 'value' >= \code{t}. If possible, one flanking row is included
        on each side of the region. If no regions achieve the specified
        \code{t}, returns NULL.
}
\description{
Given a data.frame \code{d} with 'position' and 'value' columns
output those consecutive rows where 'value' > \code{t}.
}
\details{
For objects of class \code{\link{partbl}}, \code{link{npartbl}},
\code{link{famlod}} or \code{link{fampar}}, the main data frame
containing the LOD scores per marker is used.
}
\examples{
partbl <- partbl::partbl("merlin_10_famA-parametric.tbl")
get_peaks(partbl)

}

