% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npartbl.R
\name{npartbl}
\alias{npartbl}
\title{MERLIN \code{npartbl} Object}
\usage{
npartbl(npartbl)
}
\arguments{
\item{npartbl}{The name of a \code{fam_nonparametric.tbl}
file output by MERLIN}
}
\value{
S3 object of class \code{npartbl}
}
\description{
Returns an S3 object of class \code{npartbl}
}
\examples{
npartbl_obj <- npartbl("merlin_10_famA-nonparametric.tbl")
}

