% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{partbl}
\alias{partbl}
\title{MERLIN \code{partbl} Object}
\usage{
partbl(partbl)
}
\arguments{
\item{partbl}{The name of a \code{fam_parametric.tbl}
file output by MERLIN}
}
\value{
S3 object of class \code{partbl}, which is
a list with:
\itemize{
  \item chrom: character vector of length 1
  \item n_markers: character vector of length 1
  \item gen_model: character vector of length 1
  \item partbl: data.frame with 6 columns (chr, pos, model, lod, alpha, hlod)
  }
}
\description{
Returns an S3 object of class \code{partbl} which is basically
a cleaner version of MERLIN's \code{fam_parametric.tbl} file.
}
\details{
MERLIN's \code{--tabulate} option is used to output convenient tables
summarising linkage analysis results in two files:
\code{fam_parametric.tbl} and \code{fam_nonparametric.tbl}.

The \code{fam_parametric.tbl} file will be output if MERLIN is run with
the option \code{--model genetic_model.txt}.
This uses the genetic model specified in the input file to calculate three
statistics: a multipoint parametric LOD score ('LOD' column),
an estimate of the proportion of linked families at a given locus
('ALPHA' column), and the corresponding maximum heterogeneity LOD
score ('HLOD' column).
For more information, see MERLIN's website:
\url{http://csg.sph.umich.edu/abecasis/merlin/index.html}
}
\examples{
partbl_obj <- partbl("merlin_10_famA-parametric.tbl")

}

