% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partbl.R
\name{plot.partbl}
\alias{plot.partbl}
\title{Plot MERLIN Parametric Joint LOD Scores}
\usage{
\method{plot}{partbl}(partbl, vars = c("lod", "alpha", "hlod"), ylim = c(0,
  max(partbl$max_lods[c("max_lod", "max_hlod")])))
}
\arguments{
\item{partbl}{S3 object of class \code{partbl}.}

\item{vars}{Character vector indicating which variable(s) to plot. Available
variables from the \code{partbl} data.frame are: "lod", "alpha" and "hlod".}

\item{ylim}{Character vector of length 2 indicating the range of values to plot.}
}
\value{
ggplot object, plotting the selected variables
}
\description{
Plots the LOD, ALPHA and HLOD scores in MERLIN's \code{parametric.tbl} file
}
\examples{
pt <- partbl("foo-parametric.tbl")
plot(pt)

}

