% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merlin.R
\name{plot.partbl}
\alias{plot.partbl}
\title{Plot MERLIN Parametric LOD Scores}
\usage{
\method{plot}{partbl}(partbl, vars = c("lod", "alpha", "hlod"))
}
\arguments{
\item{partbl}{An object of class \code{partbl}.}

\item{vars}{A character vector indicating which variable(s) to plot. Available
variables from the \code{partbl} file are: "lod", "alpha" and "hlod".}

\item{threshold}{Numeric of length one. Only plot values above or equal to this.}
}
\value{
A ggplot object, plotting the selected variables
}
\description{
Plots the LOD, ALPHA and HLOD scores in MERLIN's \code{famA-parametric.tbl} file
}
\examples{
pt <- partbl("foo-parametric.tbl")
plot(pt)

}

