% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npartbl.R
\name{read_merlin_npartbl}
\alias{read_merlin_npartbl}
\title{Read MERLIN \code{nonparametric.tbl}}
\usage{
read_merlin_npartbl(fname, verbose = TRUE)
}
\arguments{
\item{fname}{The file name to read.}

\item{verbose}{Display warning message for chr23.}
}
\value{
A \code{data.frame} with the columns chr, pos, analysis, lod and exlod.
}
\description{
Read the \code{nonparametric.tbl} file output by MERLIN
}
\details{
The \code{fam_nonparametric.tbl} file contains two rows at the
beginning of each 'all' and/or 'pairs' subtable, which
indicate the maximum possible scores for the specific dataset.
These rows are ignored by using the comment.char = "n"
option in \code{\link{read.table}}, which works since only these rows
contain an 'n' character. This is a bit hacky, but it should always work.
}
\examples{
read_merlin_npartbl("merlin_13_famA-nonparametric.tbl")
}

