% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partbl.R
\name{read_merlin_partbl}
\alias{read_merlin_partbl}
\title{Read MERLIN \code{parametric.tbl}}
\usage{
read_merlin_partbl(fname, verbose = TRUE)
}
\arguments{
\item{fname}{The file name to read.}

\item{verbose}{Display warning message for chr23.}
}
\value{
A \code{data.frame} with the columns chr, pos, model, lod, alpha, hlod.
}
\description{
Read the \code{parametric.tbl} file output by MERLIN.
}
\examples{
read_merlin_partbl("merlin_chr13-parametric.tbl")

}

